Clean code - Move AlertPatch mapping out of GraphPathMapper, so it can be reused from both GraphPathMapper and RaptorPath mapper - making the responsibillities clearer.
