+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/RaptorPathToItineraryMapper.java
@@ -224,7 +224,7 @@ public class RaptorPathToItineraryMapper {
         // leg.alightRule = <Assign here>;
         // leg.boardRule =  <Assign here>;
 
-        GraphPathToItineraryMapper.addAlertPatchesToLeg(
+        AlertToLegMapper.addAlertPatchesToLeg(
             request.getRoutingContext().graph,
             leg,
             firstLeg,
